+++ pkg/accesslog/events/close.go
@@ -20,6 +20,7 @@ package events
 import (
 	"time"
 
+	"github.com/apache/skywalking-rover/pkg/tools/btf/reader"
 	"github.com/apache/skywalking-rover/pkg/tools/host"
 )
 
@@ -32,6 +33,16 @@ type SocketCloseEvent struct {
 	PID       uint32
 	SocketFD  uint32
 	Success   uint32
+}
+
+func (c *SocketCloseEvent) ReadFrom(r *reader.Reader) {
+	c.ConnectionID = r.ReadUint64()
+	c.RandomID = r.ReadUint64()
+	c.StartTime = r.ReadUint64()
+	c.EndTime = r.ReadUint64()
+	c.PID = r.ReadUint32()
+	c.SocketFD = r.ReadUint32()
+	c.Success = r.ReadUint32()
 }
 
 func (c *SocketCloseEvent) GetConnectionID() uint64 {